feat: fix miner
