+++ applications/tari_miner/src/stratum/stratum_controller/controller.rs
@@ -29,8 +29,8 @@ use tari_app_grpc::tari_rpc::BlockHeader;
 use tari_utilities::hex::Hex;
 
 use crate::{
-    display_report,
     miner::Miner,
+    run_miner::display_report,
     stratum::{error::Error, stratum_types as types},
 };
