+++ integration_tests/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 
 [dependencies]
 tari_app_grpc = { path = "../applications/tari_app_grpc" }
+tari_app_utilities = { path = "../applications/tari_app_utilities" }
 tari_base_node_grpc_client = { path = "../clients/rust/base_node_grpc_client" }
 tari_wallet_grpc_client = { path = "../clients/rust/wallet_grpc_client" }
 tari_core = { path = "../base_layer/core" }
@@ -21,7 +22,7 @@ tari_console_wallet = { path = "../applications/tari_console_wallet" }
 tari_wallet = { path = "../base_layer/wallet" }
 tari_common_types = { path = "../base_layer/common_types" }
 tari_shutdown = { path = "../infrastructure/shutdown" }
-
+tari_miner = { path = "../applications/tari_miner" }
 anyhow = "1.0.53"
 async-trait = "0.1.50"
 axum = "0.6.0"
