+++ integration_tests/tests/cucumber.rs
@@ -33,15 +33,12 @@ use tari_crypto::tari_utilities::ByteArray;
 use tari_integration_tests::error::GrpcBaseNodeError;
 use tari_wallet_grpc_client::grpc::GetIdentityRequest;
 use thiserror::Error;
-use utils::{
-    miner::{mine_block_with_coinbase_on_node, mine_blocks, mine_blocks_without_wallet, register_miner_process},
-    wallet_process::spawn_wallet,
-};
+use utils::miner::{mine_block_with_coinbase_on_node, mine_blocks_without_wallet};
 
 use crate::utils::{
     base_node_process::{spawn_base_node, BaseNodeProcess},
-    miner::MinerProcess,
-    wallet_process::WalletProcess,
+    miner::{register_miner_process, MinerProcess},
+    wallet_process::{spawn_wallet, WalletProcess},
 };
 
 #[derive(Error, Debug)]
@@ -50,6 +47,8 @@ pub enum TariWorldError {
     BaseNodeProcessNotFound(String),
     #[error("Wallet process not found: {0}")]
     WalletProcessNotFound(String),
+    #[error("Miner process not found: {0}")]
+    MinerProcessNotFound(String),
     #[error("Base node error: {0}")]
     GrpcBaseNodeError(#[from] GrpcBaseNodeError),
     #[error("No base node, or wallet client found: {0}")]
@@ -73,13 +72,9 @@ enum NodeClient {
 impl TariWorld {
     async fn get_node_client<S: AsRef<str>>(
         &self,
-        name: S,
+        name: &S,
     ) -> anyhow::Result<tari_base_node_grpc_client::BaseNodeGrpcClient<tonic::transport::Channel>> {
-        self.base_nodes
-            .get(name.as_ref())
-            .ok_or_else(|| TariWorldError::BaseNodeProcessNotFound(name.as_ref().to_string()))?
-            .get_grpc_client()
-            .await
+        self.get_node(name).unwrap().get_grpc_client().await
     }
 
     async fn get_base_node_or_wallet_client<S: core::fmt::Debug + AsRef<str>>(
@@ -97,21 +92,30 @@ impl TariWorld {
 
     async fn get_wallet_client<S: AsRef<str>>(
         &self,
-        name: S,
+        name: &S,
     ) -> anyhow::Result<tari_wallet_grpc_client::WalletGrpcClient<tonic::transport::Channel>> {
-        self.wallets
-            .get(name.as_ref())
-            .ok_or_else(|| TariWorldError::WalletProcessNotFound(name.as_ref().to_string()))?
-            .get_grpc_client()
-            .await
+        self.get_wallet(name).unwrap().get_grpc_client().await
     }
 
-    #[allow(dead_code)]
-    fn get_node<S: AsRef<str>>(&self, node_name: S) -> anyhow::Result<&BaseNodeProcess> {
+    fn get_node<S: AsRef<str>>(&self, node_name: &S) -> anyhow::Result<&BaseNodeProcess> {
         Ok(self
             .base_nodes
             .get(node_name.as_ref())
             .ok_or_else(|| TariWorldError::BaseNodeProcessNotFound(node_name.as_ref().to_string()))?)
+    }
+
+    fn get_wallet<S: AsRef<str>>(&self, wallet_name: &S) -> anyhow::Result<&WalletProcess> {
+        Ok(self
+            .wallets
+            .get(wallet_name.as_ref())
+            .ok_or_else(|| TariWorldError::WalletProcessNotFound(wallet_name.as_ref().to_string()))?)
+    }
+
+    fn get_miner<S: AsRef<str>>(&self, miner_name: S) -> anyhow::Result<&MinerProcess> {
+        Ok(self
+            .miners
+            .get(miner_name.as_ref())
+            .ok_or_else(|| TariWorldError::MinerProcessNotFound(miner_name.as_ref().to_string()))?)
     }
 
     pub fn all_seed_nodes(&self) -> &[String] {
@@ -119,10 +123,10 @@ impl TariWorld {
     }
 
     pub async fn after(&mut self, _scenario: &Scenario) {
-        //     self.base_nodes.clear();
-        //     self.seed_nodes.clear();
-        //     self.wallets.clear();
-        //     self.miners.clear();
+        self.base_nodes.clear();
+        self.seed_nodes.clear();
+        self.wallets.clear();
+        self.miners.clear();
     }
 }
 
@@ -149,7 +153,7 @@ async fn start_wallet_connected_to_all_seed_nodes(world: &mut TariWorld, name: S
     spawn_wallet(world, name, None, world.all_seed_nodes().to_vec()).await;
 }
 
-#[given(expr = "a miner {word} connected to base node {word} and wallet {word}")]
+#[when(expr = "I have mining node {word} connected to base node {word} and wallet {word}")]
 async fn create_miner(world: &mut TariWorld, miner_name: String, bn_name: String, wallet_name: String) {
     register_miner_process(world, miner_name, bn_name, wallet_name);
 }
@@ -167,8 +171,8 @@ async fn node_pending_connection_to(
     first_node: String,
     second_node: String,
 ) -> anyhow::Result<()> {
-    let mut first_node = world.get_base_node_or_wallet_client(first_node).await?;
-    let mut second_node = world.get_base_node_or_wallet_client(second_node).await?;
+    let mut first_node = world.get_base_node_or_wallet_client(&first_node).await?;
+    let mut second_node = world.get_base_node_or_wallet_client(&second_node).await?;
 
     for _i in 0..100 {
         let res = match first_node {
@@ -199,7 +203,7 @@ async fn node_pending_connection_to(
 #[when(expr = "mining node {word} mines {int} blocks")]
 #[given(expr = "mining node {word} mines {int} blocks")]
 async fn run_miner(world: &mut TariWorld, miner_name: String, num_blocks: u64) {
-    mine_blocks(world, miner_name, num_blocks).await;
+    world.get_miner(miner_name).unwrap().mine(world, Some(num_blocks)).await;
 }
 
 #[then(expr = "all nodes are at height {int}")]
@@ -265,16 +269,6 @@ async fn node_is_at_height(world: &mut TariWorld, base_node: String, height: u64
         "base node didn't synchronize successfully with height {}, current chain height {}",
         height, chain_hgt
     );
-}
-
-#[when(expr = "I have mining node {word} connected to base node {word} and wallet {word}")]
-async fn miner_connected_to_base_node_and_wallet(
-    world: &mut TariWorld,
-    miner: String,
-    base_node: String,
-    wallet: String,
-) {
-    register_miner_process(world, miner, base_node, wallet);
 }
 
 #[when(expr = "I wait for wallet {word} to have at least {int} uT")]
